Auto test fix - Move shared report creation
